NPC scared state Iteration 1, Archer AI update
NPCs run away when they "Sense" someone attacking or defending. The animations might be off because the NPCs might have switched to work state before being scared. Need to set the variables for the animations.

Archer bug where the archer shoots properly when y rotation is < 180 and shoots inverted when y rotation is > 180 is fixed. Also, the arrow should rotate properly when in flight.